+++ crates/blockifier/src/fee/fee_checks.rs
@@ -1,4 +1,4 @@
-use starknet_api::execution_resources::GasAmount;
+use starknet_api::execution_resources::{GasAmount, GasVector};
 use starknet_api::transaction::Resource::{self, L1DataGas, L1Gas, L2Gas};
 use starknet_api::transaction::{AllResourceBounds, Fee, ResourceBounds, ValidResourceBounds};
 use starknet_types_core::felt::Felt;
@@ -7,7 +7,6 @@ use thiserror::Error;
 use crate::context::TransactionContext;
 use crate::fee::fee_utils::{get_balance_and_if_covers_fee, get_fee_by_gas_vector};
 use crate::fee::receipt::TransactionReceipt;
-use crate::fee::resources::GasVector;
 use crate::state::state_api::StateReader;
 use crate::transaction::errors::TransactionExecutionError;
 use crate::transaction::objects::{FeeType, TransactionExecutionResult, TransactionInfo};
