+++ crates/blockifier/src/transaction/account_transaction.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
+use starknet_api::block::GasPriceVector;
 use starknet_api::calldata;
 use starknet_api::core::{ClassHash, ContractAddress, EntryPointSelector, Nonce};
 use starknet_api::data_availability::DataAvailabilityMode;
@@ -21,7 +22,6 @@ use starknet_api::transaction::{
 use starknet_types_core::felt::Felt;
 
 use crate::abi::abi_utils::selector_from_name;
-use crate::blockifier::block::GasPricesForFeeType;
 use crate::context::{BlockContext, TransactionContext};
 use crate::execution::call_info::{CallInfo, Retdata};
 use crate::execution::contract_class::ContractClass;
@@ -328,7 +328,7 @@ impl AccountTransaction {
                         l2_gas: l2_gas_resource_bounds,
                         l1_data_gas: l1_data_gas_resource_bounds,
                     }) => {
-                        let GasPricesForFeeType { l1_gas_price, l1_data_gas_price, l2_gas_price } =
+                        let GasPriceVector { l1_gas_price, l1_data_gas_price, l2_gas_price } =
                             block_info.gas_prices.get_gas_prices_by_fee_type(fee_type);
                         vec![
                             (
