+++ crates/papyrus_execution/src/objects.rs
@@ -9,7 +9,6 @@ use blockifier::execution::call_info::{
     Retdata as BlockifierRetdata,
 };
 use blockifier::execution::entry_point::CallType as BlockifierCallType;
-use blockifier::fee::resources::GasVector;
 use blockifier::transaction::objects::{FeeType, TransactionExecutionInfo};
 use cairo_vm::types::builtin_name::BuiltinName;
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources as VmExecutionResources;
@@ -36,6 +35,7 @@ use starknet_api::deprecated_contract_class::EntryPointType;
 use starknet_api::execution_resources::{
     Builtin,
     ExecutionResources,
+    GasVector,
     GasVector as StarknetApiGasVector,
 };
 use starknet_api::state::ThinStateDiff;
